- External player support, add section to config file for <ExternalPlayers>   <ExternalPlayer>    <MediaType>Avi</MediaType>    <Command>wmplayer.exe</Command>    <Args>"{0}"</Args>   </ExternalPlayer> </ExternalPlayers>
Where ExternalPlayer element is repeated for each required mediaType. 
Valid MediaType values (case sensitive)are Avi, Mkv, DVD, BlueRay, HDDVD 
The {0} in the args parameter will be replace by the name of the item that is being played - i.e. the folder name but excluding the video_ts bit etc.

git-svn-id: https://videobrowser.googlecode.com/svn/trunk@396 096cc86f-8a4c-0410-9b85-8b5e0eee1c45